fix(Audio Playback and Muting Logic):
- Issues: Audio (music) errors occurred due to browser permissions,
whether the audio was not muted or played while it's mute.
- Changes:
  - Replaced 'musicRef' reference with 'musicActive' state because there
is a function inside useEffect for every time musicActive state change.
  - Set the initial state of 'musicActive' to null to initialize
currentTime to 0 and volume to 0.
- Added:
  - Added a global 'musicRuns' state to control music playback during
mini video animations:
      - When the mini video animation starts and 'musicRuns' is false,
'musicActive' is set to true to start the music, and 'musicRuns' is
updated to true to prevent the music from playing again through mini
video, the mini video will run the music just one time only and after
that it won't.
  - When the music button is pressed, 'musicRuns' is set to true to
prevent the music from playing through mini video when pressed.
  - If the 'isMuted' state is true, all audio is prevented from playing
until the state changes.